+++ src/dragengine/src/common/math/smooth/decSmoothDVector.h
@@ -38,7 +38,7 @@ private:
 	double pChangeSpeed;
 	double pFactorTime;
 	double pFactorLimit;
-	float pFactorDrop;
+	double pFactorDrop;
 	bool pSmoothed;
