+++ src/dragengine/src/logger/deLoggerConsole.cpp
@@ -72,7 +72,7 @@ void deLoggerConsole::LogPrefix( const char *source, const char *message, const 
 		DETHROW( deeInvalidParam );
 	}
 	
-	const int len = strlen( message );
+	const int len = ( int )strlen( message );
 	const decDateTime timestamp;
 	
 	const deMutexGuard lock( pMutex );
